biggest refactor NA
