+++ mod/node-core/pkg/components/middleware.go
@@ -29,7 +29,7 @@ import (
 	datypes "github.com/berachain/beacon-kit/mod/da/pkg/types"
 	"github.com/berachain/beacon-kit/mod/node-core/pkg/components/metrics"
 	"github.com/berachain/beacon-kit/mod/primitives"
-	"github.com/berachain/beacon-kit/mod/runtime/pkg/runtime/middleware"
+	"github.com/berachain/beacon-kit/mod/runtime/pkg/middleware"
 	depositdb "github.com/berachain/beacon-kit/mod/storage/pkg/deposit"
 )
 
@@ -101,4 +101,19 @@ func ProvideFinalizeBlockMiddleware(
 		in.ChainService,
 		in.TelemetrySink,
 	)
+}
+
+// ABCIMiddlewareInput is the input for the ABCI middleware provider.
+type ABCIMiddlewareInput struct {
+	depinject.In
+	FinalizeBlock FinalizeBlockMiddleware
+	Validator     ValidatorMiddleware
+}
+
+// ProvideABCIMiddleware is a depinject provider for the ABCI middleware.
+func ProvideABCIMiddleware(in ABCIMiddlewareInput) *ABCIMiddleware {
+	return &ABCIMiddleware{
+		FinalizeBlock: in.FinalizeBlock,
+		Validator:     in.Validator,
+	}
 }
