+++ mod/node-core/pkg/components/module/depinject.go
@@ -43,8 +43,7 @@ func init() {
 // ModuleInput is the input for the dep inject framework.
 type ModuleInput struct {
 	depinject.In
-	ABCIFinalizeBlockMiddleware components.FinalizeBlockMiddleware
-	ABCIValidatorMiddleware     components.ValidatorMiddleware
+	ABCIMiddleware *components.ABCIMiddleware
 }
 
 // ModuleOutput is the output for the dep inject framework.
@@ -57,8 +56,7 @@ type ModuleOutput struct {
 func ProvideModule(in ModuleInput) (ModuleOutput, error) {
 	return ModuleOutput{
 		Module: NewAppModule(
-			in.ABCIFinalizeBlockMiddleware,
-			in.ABCIValidatorMiddleware,
+			in.ABCIMiddleware,
 		),
 	}, nil
 }
